L2-313: CR Fixes
